DataArrayReaderを切り出しました
